refactor(app-i18n): extract data logic into hooks and use new plugin classes
