+++ packages/app-i18n/package.json
@@ -49,6 +49,7 @@
     "babel-plugin-emotion": "^9.2.8",
     "babel-plugin-named-asset-import": "^1.0.0-next.3e165448",
     "rimraf": "^3.0.2",
+    "ttypescript": "^1.5.12",
     "typescript": "^4.1.3"
   },
   "publishConfig": {
