+++ packages/app-i18n/src/admin/plugins/menus.tsx
@@ -1,13 +1,12 @@
 import React from "react";
 import { i18n } from "@webiny/app/i18n";
-import { AdminMenuPlugin } from "@webiny/app-admin/types";
-import { ReactComponent as I18NIcon } from "./../assets/icons/round-translate-24px.svg";
 import { SecureView } from "@webiny/app-security";
+import { MenuPlugin } from "@webiny/app-admin/plugins/MenuPlugin";
+import { ReactComponent as I18NIcon } from "../assets/icons/round-translate-24px.svg";
 
 const t = i18n.ns("app-form-builder/admin/menus");
 
-const plugin: AdminMenuPlugin = {
-    type: "admin-menu",
+export default new MenuPlugin({
     render({ Menu, Section, Item }) {
         return (
             <SecureView permission={"i18n.locale"}>
@@ -19,6 +18,4 @@ const plugin: AdminMenuPlugin = {
             </SecureView>
         );
     }
-};
-
-export default plugin;
+});
