+++ packages/app-i18n/src/admin/plugins/routes.tsx
@@ -2,12 +2,11 @@ import React from "react";
 import Helmet from "react-helmet";
 import { Route } from "@webiny/react-router";
 import { AdminLayout } from "@webiny/app-admin/components/AdminLayout";
-import { I18NLocales } from "../views/I18NLocales";
+import { LocalesView } from "../views/locales";
 import { SecureRoute } from "@webiny/app-security/components";
-import { RoutePlugin } from "@webiny/app/types";
+import { RoutePlugin } from "@webiny/app/plugins/RoutePlugin";
 
-const plugin: RoutePlugin = {
-    type: "route",
+export default new RoutePlugin({
     route: (
         <Route
             exact
@@ -16,12 +15,10 @@ const plugin: RoutePlugin = {
                 <SecureRoute permission={"i18n.locale"}>
                     <AdminLayout>
                         <Helmet title={"I18N - Locales"} />
-                        <I18NLocales />
+                        <LocalesView />
                     </AdminLayout>
                 </SecureRoute>
             )}
         />
     )
-};
-
-export default plugin;
+});
