--- packages/app-i18n/src/admin/views/I18NLocales/index.ts
@@ -1,1 +1,0 @@
-export { default as I18NLocales } from "./I18NLocales";
