+++ packages/app-i18n/src/admin/views/locales/hooks/useCurrentLocale.ts
@@ -1,0 +1,3 @@
+export const useCurrentLocale = () => {
+    return new URLSearchParams(location.search).get("code");
+};
