+++ packages/app-i18n/src/admin/views/locales/index.tsx
@@ -1,17 +1,17 @@
 import React from "react";
 import { SplitView, LeftPanel, RightPanel } from "@webiny/app-admin/components/SplitView";
-import I18NLocalesDataList from "./I18NLocalesDataList";
-import I18NLocalesForm from "./I18NLocalesForm";
+import LocalesDataList from "./LocalesDataList";
+import LocaleForm from "./LocaleForm";
 
-export default function I18NLocales() {
+export const LocalesView = () => {
     return (
         <SplitView>
             <LeftPanel>
-                <I18NLocalesDataList />
+                <LocalesDataList />
             </LeftPanel>
             <RightPanel>
-                <I18NLocalesForm />
+                <LocaleForm />
             </RightPanel>
         </SplitView>
     );
-}
+};
